requirements.txt
pillow>=9.0.0
numpy>=1.24
torch>=2.0.0
transformers>=4.30.0
pymilvus>=2.2.0
google-cloud-aiplatform
mysql-connector-python>=8.1.0
requests
aiohttp
urllib3